Versão Final
Modificado para ficar com aparência desejada.

+++ br.com.flordeliz/web/cliente/detalharCliente.jsp
@@ -1,9 +1,3 @@
-<%-- 
-    Document   : detalharCliente
-    Created on : 14/10/2017, 02:33:02
-    Author     : Samara C. Lima
---%>
-
 <%@page import="br.com.flordeliz.modelo.Cliente"%>
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <!DOCTYPE html>
@@ -24,8 +18,8 @@
         <p><b>Número do CPF: </b><%= cliente.getCpf() %></p>
         <p><b>Número do CNPJ da Loja: </b><%= cliente.getCnpj() %></p>
         
-        <a href="ControleCliente?acao=excluir&cliente_codigo=<%= cliente.getCodigo() %>">Excluir Cliente</a> 
-        <a href="ControleCliente?acao=atualizar&cliente_codigo=<%= cliente.getCodigo() %>">Atualizar Cliente</a><br/><br/>
+        <a class="bot_excluir" href="ControleCliente?acao=excluir&cliente_codigo=<%= cliente.getCodigo() %>" onclick="return confirm('Deseja realmente realizar a exclusão do registro?')">Excluir Cliente</a> 
+        <a class="bot_detalhar" href="ControleCliente?acao=atualizar&cliente_codigo=<%= cliente.getCodigo() %>">Atualizar Cliente</a><br/><br/>
         
     </body>
 </html>
